:bug: fix: correctly handle group status update handler
