+++ src/sentry/workflow_engine/handlers/__init__.py
@@ -2,6 +2,8 @@
 __all__ = [
     "EventCreatedByDetectorConditionHandler",
     "EventSeenCountConditionHandler",
+    "workflow_status_update_handler",
 ]
 
 from .condition import EventCreatedByDetectorConditionHandler, EventSeenCountConditionHandler
+from .workflow import workflow_status_update_handler
