+++ src/sentry/workflow_engine/handlers/workflow/__init__.py
@@ -1,0 +1,3 @@
+from .workflow_status_update_handler import workflow_status_update_handler
+
+__all__ = ["workflow_status_update_handler"]
