+++ src/sentry/workflow_engine/handlers/workflow/workflow_status_update_handler.py
@@ -1,0 +1,44 @@
+from sentry import features
+from sentry.issues.status_change_consumer import group_status_update_registry
+from sentry.issues.status_change_message import StatusChangeMessageData
+from sentry.models.activity import Activity
+from sentry.models.group import Group
+from sentry.types.activity import ActivityType
+from sentry.utils import metrics
+from sentry.workflow_engine.tasks.workflows import process_workflow_activity
+
+SUPPORTED_ACTIVITIES = [ActivityType.SET_RESOLVED.value]
+
+
+@group_status_update_registry.register("workflow_status_update")
+def workflow_status_update_handler(
+    group: Group, status_change_message: StatusChangeMessageData, activity: Activity
+) -> None:
+    """
+    Hook the process_workflow_task into the activity creation registry.
+
+    Since this handler is called in process for the activity, we want
+    to queue a task to process workflows asynchronously.
+    """
+    metrics.incr(
+        "workflow_engine.tasks.process_workflows.activity_update",
+        tags={"activity_type": activity.type},
+    )
+    if activity.type not in SUPPORTED_ACTIVITIES:
+        # If the activity type is not supported, we do not need to process it.
+        return
+
+    detector_id = status_change_message.get("detector_id")
+
+    if detector_id is None:
+        # We should not hit this case, it's should only occur if there is a bug
+        # passing it from the workflow_engine to the issue platform.
+        metrics.incr("workflow_engine.tasks.error.no_detector_id")
+        return
+
+    if features.has("organizations:workflow-engine-metric-alert-processing", group.organization):
+        process_workflow_activity.delay(
+            activity_id=activity.id,
+            group_id=group.id,
+            detector_id=detector_id,
+        )
