+++ src/sentry/workflow_engine/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .workflows import *  # noqa
