+++ tests/sentry/workflow_engine/test_task.py
@@ -9,11 +9,9 @@ from sentry.models.group import GroupStatus
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers import with_feature
 from sentry.types.activity import ActivityType
+from sentry.workflow_engine.handlers.workflow import workflow_status_update_handler
 from sentry.workflow_engine.tasks.utils import fetch_event
-from sentry.workflow_engine.tasks.workflows import (
-    process_workflow_activity,
-    workflow_status_update_handler,
-)
+from sentry.workflow_engine.tasks.workflows import process_workflow_activity
 from sentry.workflow_engine.types import WorkflowEventData
